# api_specs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Typelets API",
    "version": "1.11.5",
    "description": "A secure, encrypted notes management API with folder organization and file attachments",
    "contact": { "name": "Typelets API", "url": "https://github.com/typelets/typelets-api" }
  },
  "servers": [{ "url": "https://api.typelets.com", "description": "API Server" }],
  "components": {
    "schemas": {
...
      "Note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174001",
            "description": "Note ID"
          },
          "userId": { "type": "string", "example": "user_2abc123", "description": "User ID" },
          "folderId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Folder ID"
          },
          "title": {
            "type": "string",
            "example": "[ENCRYPTED]",
            "description": "Encrypted note title"
          },
          "content": {
            "type": "string",
            "example": "[ENCRYPTED]",
            "description": "Encrypted note content"
          },
          "encryptedTitle": {
            "type": "string",
            "nullable": true,
            "example": "base64_encrypted_data",
            "description": "Encrypted title data"
          },
          "encryptedContent": {
            "type": "string",
            "nullable": true,
            "example": "base64_encrypted_data",
            "description": "Encrypted content data"
          },
          "iv": {
            "type": "string",
            "nullable": true,
            "example": "initialization_vector",
            "description": "Initialization vector for AES-GCM"
          },
          "salt": {
            "type": "string",
            "nullable": true,
            "example": "salt_value",
            "description": "Salt for key derivation"
          },
          "starred": {
            "type": "boolean",
            "nullable": true,
            "example": false,
            "description": "Is starred"
          },
          "archived": {
            "type": "boolean",
            "nullable": true,
            "example": false,
            "description": "Is archived"
          },
          "deleted": {
            "type": "boolean",
            "nullable": true,
            "example": false,
            "description": "Is deleted (in trash)"
          },
          "hidden": {
            "type": "boolean",
            "nullable": true,
            "example": false,
            "description": "Is hidden"
          },
          "hiddenAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": null,
            "description": "When note was hidden"
          },
          "tags": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["work", "urgent"],
            "description": "Note tags"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T00:00:00.000Z",
            "description": "Created date"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-01T00:00:00.000Z",
            "description": "Updated date"
          },
          "folder": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "example": "123e4567-e89b-12d3-a456-426614174000",
                "description": "Folder ID"
              },
              "userId": { "type": "string", "example": "user_2abc123", "description": "User ID" },
              "name": { "type": "string", "example": "Work", "description": "Folder name" },
              "color": {
                "type": "string",
                "nullable": true,
                "example": "#6b7280",
                "description": "Folder color"
              },
              "parentId": {
                "type": "string",
                "nullable": true,
                "format": "uuid",
                "example": null,
                "description": "Parent folder ID"
              },
              "sortOrder": { "type": "number", "example": 0, "description": "Sort order" },
              "isDefault": {
                "type": "boolean",
                "nullable": true,
                "example": false,
                "description": "Is default folder"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "example": "2025-01-01T00:00:00.000Z",
                "description": "Created date"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "example": "2025-01-01T00:00:00.000Z",
                "description": "Updated date"
              }
            },
            "required": [
              "id",
              "userId",
              "name",
              "color",
              "parentId",
              "sortOrder",
              "isDefault",
              "createdAt",
              "updatedAt"
            ],
            "description": "Associated folder"
          }
        },
        "required": [
          "id",
          "userId",
          "folderId",
          "title",
          "content",
          "encryptedTitle",
          "encryptedContent",
          "iv",
          "salt",
          "starred",
          "archived",
          "deleted",
          "hidden",
          "hiddenAt",
          "tags",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "NotesListResponse": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/NoteWithAttachmentCount" },
            "description": "Array of notes with attachment counts"
          },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        },
        "required": ["notes", "pagination"]
      },
      "CreateNoteRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "[ENCRYPTED]",
            "description": "Must be '[ENCRYPTED]'"
          },
          "content": {
            "type": "string",
            "example": "[ENCRYPTED]",
            "description": "Must be '[ENCRYPTED]'"
          },
          "folderId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "example": null,
            "description": "Folder ID (use null or empty string for root level, or a valid folder UUID)"
          },
          "starred": { "type": "boolean", "example": false, "description": "Is starred" },
          "tags": {
            "type": "array",
            "items": { "type": "string", "maxLength": 50 },
            "maxItems": 20,
            "example": ["work"],
            "description": "Up to 20 tags, max 50 chars each"
          },
          "encryptedTitle": {
            "type": "string",
            "example": "base64_encrypted_data",
            "description": "Encrypted title"
          },
          "encryptedContent": {
            "type": "string",
            "example": "base64_encrypted_data",
            "description": "Encrypted content"
          },
          "iv": {
            "type": "string",
            "example": "initialization_vector",
            "description": "Initialization vector"
          },
          "salt": { "type": "string", "example": "salt_value", "description": "Salt value" }
        }
      },
...
    },
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk authentication token"
      }
    }
  },
  "paths": {
...
    "/api/notes/counts": {
      "get": {
        "summary": "Get note counts",
        "description": "Returns aggregated counts of notes by category. Without folder_id: returns total counts plus folders object. With folder_id: returns only a folders object (Record<folderId, counts>) with counts for each direct child folder including their descendants",
        "tags": ["Notes"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "description": "Optional. Get counts for each direct child folder of this folder ID (includes descendant notes). If omitted, returns total counts plus root-level folder counts"
            },
            "required": false,
            "description": "Optional. Get counts for each direct child folder of this folder ID (includes descendant notes). If omitted, returns total counts plus root-level folder counts",
            "name": "folder_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Note counts retrieved successfully. Without folder_id: returns {all, starred, archived, trash, folders}. With folder_id: returns Record<folderId, {all, starred, archived, trash}>",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true,
                  "example": {
                    "all": 42,
                    "starred": 5,
                    "archived": 12,
                    "trash": 3,
                    "folders": {
                      "123e4567-e89b-12d3-a456-426614174000": {
                        "all": 10,
                        "starred": 2,
                        "archived": 1,
                        "trash": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized - Invalid or missing authentication" }
        }
      }
    },
    "/api/notes": {
      "get": {
        "summary": "List notes",
        "description": "Get a paginated list of notes with optional filters for folder, starred, archived, deleted, hidden status, and search",
        "tags": ["Notes"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "description": "Filter by folder ID"
            },
            "required": false,
            "description": "Filter by folder ID",
            "name": "folderId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "example": "true",
              "description": "Filter by starred status"
            },
            "required": false,
            "description": "Filter by starred status",
            "name": "starred",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "example": "false",
              "description": "Filter by archived status"
            },
            "required": false,
            "description": "Filter by archived status",
            "name": "archived",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "example": "false",
              "description": "Filter by deleted status"
            },
            "required": false,
            "description": "Filter by deleted status",
            "name": "deleted",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["true", "false"],
              "example": "false",
              "description": "Filter by hidden status"
            },
            "required": false,
            "description": "Filter by hidden status",
            "name": "hidden",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 100,
              "example": "meeting notes",
              "description": "Search in title and content (max 100 chars, alphanumeric only)"
            },
            "required": false,
            "description": "Search in title and content (max 100 chars, alphanumeric only)",
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "example": "1",
              "description": "Page number (default: 1)"
            },
            "required": false,
            "description": "Page number (default: 1)",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 50,
              "example": "20",
              "description": "Items per page (1-50, default: 20)"
            },
            "required": false,
            "description": "Items per page (1-50, default: 20)",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Notes retrieved successfully",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotesListResponse" } }
            }
          },
          "401": { "description": "Unauthorized - Invalid or missing authentication" }
        }
      },
      "post": {
        "summary": "Create note",
        "description": "Create a new encrypted note. Title and content must be '[ENCRYPTED]' with actual encrypted data in encryptedTitle/encryptedContent fields",
        "tags": ["Notes"],
        "security": [{ "Bearer": [] }],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateNoteRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Note created successfully",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Note" } } }
          },
          "400": { "description": "Invalid request body" },
          "401": { "description": "Unauthorized - Invalid or missing authentication" },
          "402": { "description": "Payment required - Note limit exceeded" }
        }
      }
    },
...
    "/api/folders": {
      "get": {
        "summary": "List folders",
        "description": "Returns a paginated list of folders. Supports filtering by parent folder. Results include note counts and child folders.",
        "tags": ["Folders"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "description": "Filter by parent folder ID (omit for root-level folders)"
            },
            "required": false,
            "description": "Filter by parent folder ID (omit for root-level folders)",
            "name": "parentId",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "example": "1",
              "description": "Page number (default: 1)"
            },
            "required": false,
            "description": "Page number (default: 1)",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 50,
              "example": "20",
              "description": "Items per page (1-50, default: 20)"
            },
            "required": false,
            "description": "Items per page (1-50, default: 20)",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Folders retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FoldersListResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized - Invalid or missing authentication" }
        }
      },
      "post": {
        "summary": "Create folder",
        "description": "Creates a new folder. Optionally specify a parent folder for nested organization.",
        "tags": ["Folders"],
        "security": [{ "Bearer": [] }],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateFolderRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Folder created successfully",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Folder" } }
            }
          },
          "400": { "description": "Bad request - Invalid input or parent folder not found" },
          "401": { "description": "Unauthorized - Invalid or missing authentication" }
        }
      }
    },
...
  }
}